generative_ai/pyproject.toml
[tool.poetry]
name = "generative_ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
openai-functions-agent = {path = "packages\\openai-functions-agent", develop = true}
langchain-pinecone = "^0.1.1"
lxml = "^5.2.2"
bs4 = "^0.0.2"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
